adgm_corporate_agent/requirements.txt
gradio
streamlit
python-docx
langchain
langchain-community
faiss-cpu
beautifulsoup4
requests
pypdf
tiktoken
ollama